Create bounding boxes for countours of detected rings
I updated the TestThresholding class to get the contours of the thresholded frames
(now adjusted to detect rings) and draw a bounding box around it if it was large enough.
I also updated the ThresholdInRange class to show the exact H, S, or V values selected in
order to facilitate tuning the threshold values of TestThresholding.